% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_masonry.R
\name{masonry_effect_list}
\alias{masonry_effect_list}
\title{Return a list of all  Masonry effects}
\usage{
masonry_effect_list()
}
\description{
Return a list of all  Masonry effects
}
\examples{
masonry_effect_list()

}
