% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_badge.R
\name{material_badge}
\alias{material_badge}
\title{Create a badge}
\usage{
material_badge(content, type = NULL, color = NULL, custom_caption = FALSE)
}
\arguments{
\item{content}{Badge content.}

\item{type}{Badge type. NULL by default, can be set as "new".}

\item{color}{Badge color. NULL by default. Visit http://materializecss.com/color.html for a list of available colors.}

\item{custom_caption}{Whether to enable custom captions or not. FALSE by default.}
}
\description{
UI content that can be placed everywhere
}
\examples{
material_badge(
 content = "1",
 type = "new"
)
}
