% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{material_blockquote}
\alias{material_blockquote}
\title{Create a material blockquote}
\usage{
material_blockquote(...)
}
\arguments{
\item{...}{Any element, such as text.}
}
\description{
Create a material blockquote
}
\examples{
material_blockquote(
 "This is an example quotation that uses the blockquote tag."
)
}
