% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_blog_card.R
\name{material_blog_card}
\alias{material_blog_card}
\title{Create blog card}
\usage{
material_blog_card(..., src = NULL, author_name = NULL, post_title = NULL,
  date = NULL, media = NULL, extra = NULL, icon = "info")
}
\arguments{
\item{...}{Any element}

\item{src}{Post author image url or path.}

\item{author_name}{Post author.}

\item{post_title}{Post title.}

\item{date}{Date of publication.}

\item{media}{Media content. Can be an image and/or video.}

\item{extra}{Extra content, if any. Displayed below the media.}

\item{icon}{Post icon, if any.}
}
\description{
Create blog card
}
\note{
Be careful: all video should have the embed attribut such as https://www.youtube.com/embed/10r9ozshGVE.
Otherwise they will no render.
}
\examples{
material_blog_card(
 src = "https://pixinvent.com/materialize-material-design-admin-template/images/avatar/avatar-7.png",
 author_name = "Jean Eude",
 post_title = "My Post",
 date = "14 May 2018",
 media = material_media(
  src = "https://pixinvent.com/materialize-material-design-admin-template/images/gallary/3.png"
 )
),
 "Le Lorem Ipsum est simplement du faux texte employé dans la 
 composition et la mise en page avant impression. Le Lorem 
 Ipsum est le faux texte standard de l'imprimerie depuis les 
 années 1500, quand un peintre anonyme assembla ensemble des 
 morceaux de texte pour réaliser un livre spécimen de polices 
 de texte. Il n'a pas fait que survivre cinq siècles, mais s'est 
 aussi adapté à la bureautique informatique, sans que son contenu 
 n'en soit modifié. Il a été popularisé dans les années 1960 grâce 
 à la vente de feuilles Letraset contenant des passages du Lorem 
 Ipsum, et, plus récemment, par son inclusion dans des applications 
 de mise en page de texte, comme Aldus PageMaker"
)
}
