% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-card.R
\name{material_card}
\alias{material_card}
\title{Create a card that will contain UI content}
\usage{
material_card(title, ..., size = "medium", depth = NULL, image = FALSE,
  src = NULL, activator = FALSE, tabs = FALSE, tabscontent = NULL,
  extra = NULL, color = NULL, hoverable = TRUE)
}
\arguments{
\item{title}{String. The title of the card.}

\item{...}{The UI elements to place in the card.}

\item{size}{Card size: "small", "medium", "large". "medium" by default.}

\item{depth}{Integer. The amount of depth of the card. The value should be between 0 and 5. Leave empty for the default depth.}

\item{image}{Whether to enable image in the card. FALSE by default.}

\item{src}{Image path, if any.}

\item{activator}{Whether to allow the card reveal. FALSE by default.}

\item{tabs}{Whether to display tabs in the card. FALSE by default.}

\item{tabscontent}{Tabs content, if any.}

\item{extra}{Card reveal content if any.}

\item{color}{String. The card color. Leave empty for the default color. Visit http://materializecss.com/color.html for a list of available colors.}

\item{hoverable}{Whether to hover the card. TRUE by default.}
}
\description{
UI content can be placed in cards to organize items on a page.
}
\examples{
material_card(
  title = "Example Card",
  depth = 5,
  shiny::tags$h5("Card Content")
)
}
