% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-card.R
\name{material_card_horizontal}
\alias{material_card_horizontal}
\title{Create a card that is horizontally displayed}
\usage{
material_card_horizontal(..., title = NULL, src = NULL, depth = NULL,
  hoverable = FALSE)
}
\arguments{
\item{...}{Any element.}

\item{title}{Card Title.}

\item{src}{Image path, if any.}

\item{depth}{Integer. The amount of depth of the card. The value should be between 0 and 5. Leave empty for the default depth.}

\item{hoverable}{Whether to hover the card. TRUE by default.}
}
\description{
UI content can be placed in cards to organize items on a page.
}
\note{
Render well if embeded in a material_column of width 12.
}
\examples{
material_card_horizontal(
  src = "https://pixinvent.com/materialize-material-design-admin-template/images/gallary/21.png",
  title = "Card Title",
  "It is a long established fact that a reader will be distracted 
  by the readable content of a page when looking at its layout. 
  The point of using Lorem Ipsum is that it has a more-or-less 
  normal distribution of letters, as opposed to using 'Content here, 
  Packages and web page editors now use Lorem Ipsum as their default 
  model text, and a search for 'lorem ipsum' will uncover. 
  Web page editors now use Lorem Ipsum as their default"
)
}
