% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_card_tabs.R
\name{material_card_tabs}
\alias{material_card_tabs}
\title{Create a slot for a card tab content}
\usage{
material_card_tabs(..., id)
}
\arguments{
\item{...}{Slot for any UI element.}

\item{id}{Tag id. Should be unique and correspond to the id provided in material_tabs.}
}
\description{
UI content can be placed in cards.
}
\examples{
material_card_tabs(
 id = "tab_1"
 "Content"
)
}
