% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material-medias.R
\name{material_carousel}
\alias{material_carousel}
\title{Create a material carousel}
\usage{
material_carousel(..., slider = FALSE, slider_position = NULL)
}
\arguments{
\item{...}{slot for material_carousel_item.}

\item{slider}{Whether to display a slider or not. FALSE by default.}

\item{slider_position}{Slider position, only if slider is TRUE.}
}
\description{
Create a material carousel
}
\examples{
material_carousel(
 material_carousel_item(
  src = "https://cdn4.buysellads.net/uu/1/3386/1525211184-62491.png"
 ),
 material_carousel_item(
  fixed = TRUE,
  src = "https://cdn4.buysellads.net/uu/1/3386/1525211184-62491.png"
 )
)
}
