% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_carousel.R
\name{material_carousel_item}
\alias{material_carousel_item}
\title{Create a material carousel item}
\usage{
material_carousel_item(..., src = NULL, enable_content = FALSE,
  title = NULL, color = NULL, fixed = FALSE)
}
\arguments{
\item{...}{any element. Only if enable_content is TRUE.}

\item{src}{Image path or url.}

\item{enable_content}{Whether to enable content or not. FALSE by default.}

\item{title}{Item title. Only if enable_content is TRUE.}

\item{color}{Item color. Only available when enable_content is TRUE. Visit http://materializecss.com/color.html for a list of available colors.}

\item{fixed}{Whether to fix the content or not. FALSE by default.}
}
\description{
Create a material carousel item
}
