% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{material_collapsible}
\alias{material_collapsible}
\title{Create a collapsible element}
\usage{
material_collapsible(..., type = NULL)
}
\arguments{
\item{...}{slot for material_collapsible_item.}

\item{type}{Collapse behavior. NULL by default (accordion). Can also be "expandable" or "popout".}
}
\description{
Create a collapsible element
}
\note{
expandable does not work at the moment.
}
\examples{
material_collapsible(
 material_collapsible_item(
  "Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
  sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. 
  Ut enim ad minim veniam, quis nostrud exercitation ullamco 
  laboris nisi ut aliquip ex ea commodo consequat",
  header = "First",
  active = TRUE
 ),
 material_collapsible_item(
  "Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
  sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. 
  Ut enim ad minim veniam, quis nostrud exercitation ullamco 
  laboris nisi ut aliquip ex ea commodo consequat",
  header = "Second"
 ),
 material_collapsible_item(
  "Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
  sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. 
  Ut enim ad minim veniam, quis nostrud exercitation ullamco 
  laboris nisi ut aliquip ex ea commodo consequat",
  header = "Third"
 )
)
}
