% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_collapsible.R
\name{material_collapsible_item}
\alias{material_collapsible_item}
\title{Create a collapsible item}
\usage{
material_collapsible_item(..., header = NULL, icon = NULL, active = FALSE)
}
\arguments{
\item{...}{Body content.}

\item{header}{Item title.}

\item{icon}{Item icon, if any.}

\item{active}{Whether the item is active or not. FALSE by default.}
}
\description{
Create a collapsible item
}
