% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_collection.R
\name{material_collection}
\alias{material_collection}
\title{Create a collection}
\usage{
material_collection(..., header = FALSE, depth = NULL)
}
\arguments{
\item{...}{slot for material_collection_item or material_collection_header or material_collection_avatar.}

\item{header}{Whether to enable material_collection_header. FALSE by default.}

\item{depth}{Collection depth. Value between 0 and 5.}
}
\description{
Create a collection
}
\examples{
material_collection(
  header = TRUE,
  material_collection_header(
   title = "item 1"
  ),
  material_collection_item(
   active = TRUE,
   "This is the item content",
   secondary_content = material_badge(
    content = "Secondary content", 
    color = "red"
   )
  ),
  material_collection_avatar(
   src = "https://materializecss.com/images/yuna.jpg",
   title = "Title",
   content = "This is the content"
  )
)
}
