% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{material_collection_avatar}
\alias{material_collection_avatar}
\title{Create a collection avatar}
\usage{
material_collection_avatar(src = NULL, title = NULL, content = NULL,
  secondary_content = NULL)
}
\arguments{
\item{src}{Avatar image path or url.}

\item{title}{Avatar title.}

\item{content}{Content, if any.}

\item{secondary_content}{Secondary content if any. Displayed on the right side.}
}
\description{
Create a collection avatar
}
