% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{material_collection_header}
\alias{material_collection_header}
\title{Create a collection header}
\usage{
material_collection_header(title = NULL)
}
\arguments{
\item{title}{Header content.}
}
\description{
Create a collection header
}
