% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{material_collection_item}
\alias{material_collection_item}
\title{Create a collection item}
\usage{
material_collection_item(..., active = FALSE, secondary_content = NULL)
}
\arguments{
\item{...}{item content.}

\item{active}{Whether the item is active or not. FALSE by default.}

\item{secondary_content}{Secondary content if any. Displayed on the right side.}
}
\description{
Create a collection item
}
