% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{material_column}
\alias{material_column}
\title{Create a column to organize UI content}
\usage{
material_column(..., s_width = 12, m_width = 6, l_width = 6,
  xl_width = 6, offset = 0)
}
\arguments{
\item{...}{The UI elements to place in the column.}

\item{s_width}{Integer. The width of the column on mobile devices (<= 600px). The value should be between 1 and 12.}

\item{m_width}{Integer. The width of the column on tablet devices (> 600px). The value should be between 1 and 12.}

\item{l_width}{Integer. The width of the column on desktop devices (> 992px). The value should be between 1 and 12.}

\item{xl_width}{Integer. The width of the column on large desktop devices (> 1200px). The value should be between 1 and 12.}

\item{offset}{Integer. The offset to the left of the column. The value should be between 0 and 11.}
}
\description{
UI content can be placed in columns to organize items on a page.
}
\examples{
material_column(
  shiny::tags$h1("Column Content")
)
}
