% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typography.R
\name{material_container}
\alias{material_container}
\title{Create a material container div}
\usage{
material_container(...)
}
\arguments{
\item{...}{Any UI element.}
}
\description{
To center elements inside
}
\examples{
material_container(
 material_card(
  title = "Example Card",
  depth = 5,
  shiny::tags$h5("Card Content")
 )
)
}
