% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_discovery.R
\name{material_discovery}
\alias{material_discovery}
\title{Create a material discovery}
\usage{
material_discovery(id, title = NULL, description = NULL)
}
\arguments{
\item{id}{Discovery id. Should be unique.}

\item{title}{Discovery title.}

\item{description}{DIscovery description.}
}
\description{
Create a material discovery
}
\examples{

material_discovery_trigger()
material_discovery(
 id = "discovery_button",
 title = "My discovery",
 description = "A bunch of text"
)
}
