% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_discovery.R
\name{material_discovery_trigger}
\alias{material_discovery_trigger}
\title{Create a material discovery trigger}
\usage{
material_discovery_trigger(label = NULL)
}
\arguments{
\item{label}{Discovery trigger label.}
}
\description{
Create a material discovery trigger
}
