% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_divider.R
\name{material_divider}
\alias{material_divider}
\title{Create an horizontal divider}
\usage{
material_divider()
}
\description{
Create an horizontal divider
}
\note{
Perfect to insert in a card.
}
\examples{
material_divider()

}
