% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_footer.R
\name{material_footer}
\alias{material_footer}
\title{Create a shinymaterial footer}
\usage{
material_footer(..., footer_color = NULL, copyright = NULL,
  footer_extra = NULL)
}
\arguments{
\item{...}{Footer content.}

\item{footer_color}{Footer color.}

\item{copyright}{Copyright text in footer.}

\item{footer_extra}{Display additional text on the right of copyrights.}
}
\description{
Build a shinymaterial footer
}
