% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_gallery.R
\name{material_gallery}
\alias{material_gallery}
\title{Create a material gallery}
\usage{
material_gallery(...)
}
\arguments{
\item{...}{Slot for material_gallery_item.}
}
\description{
Create a material gallery
}
\examples{
material_gallery(
 material_gallery_item(
  src = "https://pixinvent.com/materialize-material-design-admin-template/images/gallary/22.png",
  "A beautiful image"
 ),
 material_gallery_item(
  src = "https://cdn.dribbble.com/users/329207/screenshots/3761869/bemocs_rei_labor_day_dribbble.jpg",
  "Another beautiful image"
 ),
 material_gallery_item(
  src = "https://img00.deviantart.net/2cb5/i/2017/048/a/4/blue_mountain_by_jordangrimmer-dazetlq.jpg",
  "Yet another beautiful image"
 )
)
}
