% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material-medias.R
\name{material_gallery_item}
\alias{material_gallery_item}
\title{Create a material gallery item}
\usage{
material_gallery_item(..., src = NULL)
}
\arguments{
\item{...}{Slot for any UI element.}

\item{src}{Image path or url, if any.}
}
\description{
Create a material gallery item
}
