% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_gradient_card.R
\name{material_gradient_card}
\alias{material_gradient_card}
\title{Create a material gradient card}
\usage{
material_gradient_card(title = NULL, subtitle = NULL, src = NULL,
  color = NULL, shadow = TRUE)
}
\arguments{
\item{title}{Gradient card title.}

\item{subtitle}{Gradient card subtitle.}

\item{src}{Gradient card image url or path.}

\item{color}{Card color. Use the function material_gradient_color() to have the full list.}

\item{shadow}{Card gradient shadow. TRUE by default.}
}
\description{
Create a material gradient card
}
\examples{
material_gradient_card(
 src = "https://pixinvent.com/materialize-material-design-admin-template/images/icon/apple-watch.png",
 title = "50\% Off",
 subtitle = "On apple watch",
 color = material_gradient_color()[[15]]
)
}
