% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_info_card.R
\name{material_gradient_color}
\alias{material_gradient_color}
\title{Return a material gradient list}
\usage{
material_gradient_color()
}
\description{
Return a material gradient list
}
\examples{
material_gradient_color()

}
