% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_info_card.R
\name{material_info_card}
\alias{material_info_card}
\title{Create a material info card}
\usage{
material_info_card(icon = NULL, title = NULL, value = NULL,
  subtitle = NULL, extra = NULL, color = NULL, shadow = TRUE)
}
\arguments{
\item{icon}{Info card icon.}

\item{title}{Info card title.}

\item{value}{Info card value.}

\item{subtitle}{Info card subtitle.}

\item{extra}{info card extra information.}

\item{color}{Card color. Use the function material_gradient_color() to have the full list.}

\item{shadow}{Card gradient shadow. TRUE by default.}
}
\description{
Create a material info card
}
\examples{
material_info_card(
 icon = "timeline",
 title = "Orders",
 value = 690,
 subtitle = "New",
 extra = 6000,
 color = material_gradient_color()[[15]]
)
}
