% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_masonry.R
\name{material_masonry}
\alias{material_masonry}
\title{Create a material Masonry media}
\usage{
material_masonry(effect, src = NULL, caption_1 = NULL, caption_2 = NULL,
  subcaption = NULL)
}
\arguments{
\item{effect}{Masonry effect: choose among the list given in masonry_effect_list().}

\item{src}{Image url or path.}

\item{caption_1}{First part of title.}

\item{caption_2}{Second part of title.}

\item{subcaption}{Subtitle.}
}
\description{
Create a material Masonry media
}
\examples{
material_masonry(
 effect = masonry_effect_list()[[4]],
 src = "https://pixinvent.com/materialize-material-design-admin-template/images/gallary/5.png",
 caption_1 = "Nice",
 caption_2 = "Lily",
 subcaption = "Lily likes to play with crayons and pencils"
)
}
