% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_media.R
\name{material_media}
\alias{material_media}
\title{Create a material media container}
\usage{
material_media(src = NULL, width = NULL)
}
\arguments{
\item{src}{Media path or url, if any.}

\item{width}{media width in px. Should be NULL if responsive is TRUE.}
}
\description{
Create a material media container
}
\examples{
material_media(
 src = "https://materializecss.com/images/sample-1.jpg",
 width = "650"
)
}
