% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_navbar.R
\name{material_navbar}
\alias{material_navbar}
\title{Create a shinymaterial footer}
\usage{
material_navbar(title = "", nav_bar_fixed = FALSE,
  nav_bar_color = NULL, font_color = NULL)
}
\arguments{
\item{title}{String. The title of the page.}

\item{nav_bar_fixed}{Boolean. Should the nav bar remain fixed on the screen?}

\item{nav_bar_color}{Color of the nav bar. Leave blank for the default color. 
Visit \url{http://materializecss.com/color.html} for a list of available colors. 
 @param font_color String. The title font color. Leave blank for the default color. 
Visit \url{http://materializecss.com/color.html} for a list of available colors. 
\emph{Title color requires using word forms of colors (e.g. "deep-purple"). 
Also, lighten or darken effects do not work on title colors.}}
}
\description{
Build a shinymaterial footer
}
