% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-material-page.R
\name{material_page}
\alias{material_page}
\title{Create a shinymaterial page}
\usage{
material_page(..., enable_footer = TRUE, navbar = NULL,
  footer = NULL, background_color = "grey lighten-4",
  include_fonts = FALSE)
}
\arguments{
\item{...}{The UI elements to place in the page.}

\item{enable_footer}{Whether to display the footer or not. TRUE by default.}

\item{navbar}{Slot for \link{material_navbar}.}

\item{footer}{Slot for \link{material_footer}.}

\item{background_color}{Page background color. Leave blank for the default color. 
Visit \url{http://materializecss.com/color.html} for a list of available colors.}

\item{include_fonts}{Boolean. Should the material font files be included? (This will place the font sources in a directory 'www', at the same location as the app code.)}
}
\description{
Build a shinymaterial page.
}
\examples{
material_page(
  material_navbar(
   
  )
  background_color = "blue lighten-4",
  shiny::tags$h1("Page Content")
)
}
