% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_plans.R
\name{material_plan_collection}
\alias{material_plan_collection}
\title{Create a material plan collection}
\usage{
material_plan_collection(...)
}
\arguments{
\item{...}{Slot for material_plan_collection_item}
}
\description{
Create a material plan collection
}
