% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_plans.R
\name{material_plan_collection_item}
\alias{material_plan_collection_item}
\title{Create a material plan collection item}
\usage{
material_plan_collection_item(description = NULL)
}
\arguments{
\item{description}{Item description.}
}
\description{
Create a material plan collection item
}
