% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_plans.R
\name{material_plan_item}
\alias{material_plan_item}
\title{Create a material plan item}
\usage{
material_plan_item(..., nb_plan = 3, title = NULL, value = NULL,
  subtitle = NULL, url = NULL, depth = NULL, color = NULL)
}
\arguments{
\item{...}{Slot for material_plan_collection}

\item{nb_plan}{Number of material_plan_item.}

\item{title}{Item title.}

\item{value}{Item subtitle.}

\item{subtitle}{Item subtitle.}

\item{url}{Item external link.}

\item{depth}{Item depth. Between 0 and 5. NULL by default.}

\item{color}{Item color. Use the function material_gradient_color() to have the full list.}
}
\description{
Create a material plan item
}
