% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_plans.R
\name{material_plans}
\alias{material_plans}
\title{Create a material plan}
\usage{
material_plans(...)
}
\arguments{
\item{...}{Slot for material_plan_item.}
}
\description{
Create a material plan
}
\examples{
material_plans(
 nb_plan = 3,
 material_plan_item(
  title = "Basic",
  value = "10$",
  subtitle = "Free 1 month",
  url = "http://www.google.com",
  depth = 5,
  color = material_gradient_color()[[15]],
  material_plan_collection(
   material_plan_collection_item(
   description = "500 emails"
   ),
   material_plan_collection_item(
   description = "Unlimited data"
   ),
   material_plan_collection_item(
   description = "1 user"
   )
  )
 ),
 material_plan_item(
  title = "Pro",
  value = "29$",
  subtitle = "Most popular",
  url = "http://www.google.com",
  depth = NULL,
  color = material_gradient_color()[[4]],
  material_plan_collection(
   material_plan_collection_item(
   description = "2000 emails"
   ),
   material_plan_collection_item(
   description = "Unlimited data"
   ),
   material_plan_collection_item(
   description = "10 users"
   )
  )
 ),
 material_plan_item(
  title = "Premium",
  value = "49$",
  subtitle = "Get 20\% off",
  url = "http://www.google.com",
  depth = NULL,
  color = material_gradient_color()[[20]],
  material_plan_collection(
   material_plan_collection_item(
   description = "10000 emails"
   ),
   material_plan_collection_item(
   description = "Unlimited data"
   ),
   material_plan_collection_item(
   description = "Unlimited user"
   )
  )
 )
)
}
