% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_progress.R
\name{material_progress}
\alias{material_progress}
\title{Create a material progress bar}
\usage{
material_progress(main_color = NULL, value = "50", secondary_color = NULL)
}
\arguments{
\item{main_color}{Progress color. NULL by default. Visit http://materializecss.com/color.html for a list of available colors.}

\item{value}{Progress value (between 0 and 100)}

\item{secondary_color}{Progress secondary color. Represents the 100-value part. NULL by default. Visit http://materializecss.com/color.html for a list of available colors.}
}
\description{
Create a material progress bar
}
\examples{
material_progress(
 value = "70\%"
)
}
