% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_right_nav.R
\name{material_right_nav}
\alias{material_right_nav}
\title{Create a shinymaterial right sidebar}
\usage{
material_right_nav()
}
\description{
Create a shinymaterial right sidebar
}
\examples{
material_right_nav(
  
)
}
