% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_right_nav.R
\name{material_right_nav_TabContent}
\alias{material_right_nav_TabContent}
\title{Create a shinymaterial right sidebar tab content}
\usage{
material_right_nav_TabContent(..., id, tabName = NULL)
}
\arguments{
\item{...}{Any UI element, such as sliders, checkboxes, text, ...}

\item{id}{Unique item id. Should correspond to the id given in material_right_nav_TabItem.}

\item{tabName}{Tab name.}
}
\description{
Create a shinymaterial right sidebar tab content
}
