% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_right_nav.R
\name{material_right_nav_TabItem}
\alias{material_right_nav_TabItem}
\title{Create a shinymaterial right sidebar tab list item}
\usage{
material_right_nav_TabItem(id, tabIcon = NULL)
}
\arguments{
\item{id}{Unique item id. String.}

\item{tabIcon}{Tab icon.}
}
\description{
Create a shinymaterial right sidebar tab list item
}
