% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_right_nav.R
\name{material_right_nav_TabList}
\alias{material_right_nav_TabList}
\title{Create a shinymaterial right sidebar tab list}
\usage{
material_right_nav_TabList(...)
}
\arguments{
\item{...}{Slot for material_right_nav_TabItem.}
}
\description{
Create a shinymaterial right sidebar tab list
}
