% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_toc.R
\name{material_scrollpsy}
\alias{material_scrollpsy}
\title{Create a material scrollpsy}
\usage{
material_scrollpsy(...)
}
\arguments{
\item{...}{Slot for material_scrollpsy_item.}
}
\description{
Create a material scrollpsy
}
\examples{
material_scrollpsy(
 material_scrollpsy_item(target = "part1"),
 material_scrollpsy_item(target = "part2"),
 material_scrollpsy_item(target = "part3")
)
}
