% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_toc.R
\name{material_scrollpsy_item}
\alias{material_scrollpsy_item}
\title{Create a material table of content item}
\usage{
material_scrollpsy_item(target = NULL)
}
\arguments{
\item{target}{id element to target. The target should be wrap in a div(id = ..., ).}
}
\description{
Create a material table of content item
}
