% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{material_scrollspy}
\alias{material_scrollspy}
\title{Create a material scrollspy}
\usage{
material_scrollspy(...)
}
\arguments{
\item{...}{Slot for material_scrollspy_item.}
}
\description{
Create a material scrollspy
}
\examples{
material_scrollspy(
 material_scrollspy_item(target = "part1"),
 material_scrollspy_item(target = "part2"),
 material_scrollspy_item(target = "part3")
)
}
