% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{material_scrollspy_item}
\alias{material_scrollspy_item}
\title{Create a material scrollspy item}
\usage{
material_scrollspy_item(target = NULL)
}
\arguments{
\item{target}{id element to target. The target should be wrap in a div(id = ..., ).}
}
\description{
Create a material scrollspy item
}
