% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_stat_card.R
\name{material_stat_card}
\alias{material_stat_card}
\title{Create a material stat card}
\usage{
material_stat_card(title = NULL, value = NULL, trend = NULL,
  subtitle = NULL, footer = NULL, icon = NULL, color = NULL)
}
\arguments{
\item{title}{Card title.}

\item{value}{Card subtitle.}

\item{trend}{Card trend. "more", "less" or "equal".}

\item{subtitle}{Card subtitle.}

\item{footer}{Card footer, if any.}

\item{color}{Card color. Use the function material_gradient_color() to have the full list.}
}
\description{
Create a material stat card
}
\examples{
material_stat_card(
 title = "New Clients",
 value = 566,
 trend = "up",
 subtitle = "15\% from yesterday",
 color = "green"
)
}
