% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_task_card.R
\name{material_task_card}
\alias{material_task_card}
\title{Create a material task card}
\usage{
material_task_card(..., title = NULL, date = NULL, color = NULL,
  depth = NULL, hoverable = FALSE)
}
\arguments{
\item{...}{Slot for material_task_item.}

\item{title}{Task card title.}

\item{date}{Task card date.}

\item{color}{Task header color.}

\item{depth}{Integer. The amount of depth of the card. The value should be between 0 and 5. Leave empty for the default depth.}

\item{hoverable}{Whether to hover the card. TRUE by default.}
}
\description{
Create a material task card
}
\examples{
material_task_card(
  title = "Task Card",
  date = "14 may 2018",
  color = "cyan",
  material_task_item(
   id = 1, 
   name = "task 1", 
   date = "today",
   tag_name = "tag", 
   tag_color = "red",
   checked = FALSE, 
   disabled = FALSE
  ),
  material_task_item(
   id = 2, 
   name = "task 2", 
   date = "today",
   tag_name = "tag", 
   tag_color = "orange",
   checked = TRUE, 
   disabled = FALSE
  )
)
}
