% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_task_card.R
\name{material_task_item}
\alias{material_task_item}
\title{Create a material task item}
\usage{
material_task_item(id, name = NULL, date = NULL, tag_name = NULL,
  tag_color = NULL, checked = FALSE, disabled = FALSE)
}
\arguments{
\item{id}{Task id. Should be unique.}

\item{name}{Task name.}

\item{date}{Task date.}

\item{tag_name}{Task tag name.}

\item{tag_color}{Taks tag color.}

\item{checked}{Whether the task is checked or not at start.}

\item{disabled}{Whether the task is disabled. Useful for old tasks.}
}
\description{
Create a material task item
}
