% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_timeline.R
\name{material_timeline}
\alias{material_timeline}
\title{Create a material timeline}
\usage{
material_timeline(...)
}
\arguments{
\item{...}{Slot for material_timeline_item}
}
\description{
Create a material timeline
}
\examples{
material_timeline(
 material_timeline_item(
  material_timeline_card(
   src = "https://i.pinimg.com/originals/df/0a/3e/df0a3e2ec30abb1c92d145ef165b714f.gif",
   depth = 5,
   title = "Item 1",
   extra = "Some text here!"
  ),
  marker_icon = "language",
  marker_color = "blue",
  marker_size = "large"
 ),
 material_timeline_item(
  material_timeline_card(
   src = "https://kulturologia.ru/files/u12645/Rhads-2.jpg",
   hoverable = TRUE,
   title = "Item 2",
   extra = "Some text here!"
  ),
  marker_icon = "work",
  marker_color = "green",
  marker_size = "small"
 )
)
}
