% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_timeline.R
\name{material_timeline_card}
\alias{material_timeline_card}
\title{Create a material timeline card}
\usage{
material_timeline_card(..., src = NULL, title = NULL, extra = NULL,
  depth = NULL, hoverable = TRUE)
}
\arguments{
\item{...}{Card content}

\item{src}{Card image path or url.}

\item{title}{Card title.}

\item{extra}{Card reveal extra.}

\item{depth}{Integer. The amount of depth of the card. The value should be between 0 and 5. Leave empty for the default depth.}

\item{hoverable}{Whether to hover the card. TRUE by default.}
}
\description{
Create a material timeline card
}
