% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_timeline.R
\name{material_timeline_item}
\alias{material_timeline_item}
\title{Create a material timeline item}
\usage{
material_timeline_item(..., marker_icon = NULL, marker_color = NULL,
  marker_size = NULL)
}
\arguments{
\item{...}{Any element such as material_timeline_card.}

\item{marker_icon}{Marker icon.}

\item{marker_color}{Marker color.}

\item{marker_size}{Marker size.}
}
\description{
Create a material timeline item
}
