% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material-medias.R
\name{material_video}
\alias{material_video}
\title{Create a material video container}
\usage{
material_video(src = NULL, width = NULL, height = NULL,
  responsive = TRUE, fullscreen = TRUE)
}
\arguments{
\item{src}{Video path or url, if any.}

\item{width}{Video width in px. Should be NULL if responsive is TRUE.}

\item{height}{Video height in px. Should be NULL if responsive is TRUE.}

\item{responsive}{Whether the video is responsive. TRUE by default.}

\item{fullscreen}{Whether to allow full screen. Useless if responsive is TRUE.}
}
\description{
Create a material video container
}
\examples{
material_video(
 src = "//www.youtube.com/embed/Q8TXgCzxEnw?rel=0",
 responsive = TRUE
)
}
