% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_chip.R
\name{material_chip}
\alias{material_chip}
\title{Create a material badge (chip)}
\usage{
material_chip(..., src = NULL, chip_name = NULL)
}
\arguments{
\item{src}{Image path or url, if any.}

\item{chip_name}{Chip name.}

\item{Any}{element such as material_icon...}
}
\description{
Create a material badge (chip)
}
\examples{
material_chip(
 src = "https://pixinvent.com/materialize-material-design-admin-template/images/avatar/avatar-1.png",
 chip_name = "John Doe",
 material_icon("close")
)
}
