% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful-items.R
\name{material_icon}
\alias{material_icon}
\title{Create a material icon}
\usage{
material_icon(name, circle = FALSE, color = NULL, position = NULL,
  closable = FALSE, size = NULL)
}
\arguments{
\item{name}{Icon to display. Choose here: https://materializecss.com/icons.html.}

\item{circle}{Whether to circle the icon. FALSE by default.}

\item{color}{Icon color. Use when circle is TRUE. Visit http://materializecss.com/color.html for a list of available colors.}

\item{position}{Icon position. NULL, "right" or "left".}

\item{closable}{Whether the icon is closable or not. Perfect to be insert in material_chip.}

\item{size}{Icon size: "tiny", "small", "medium" or "large".}
}
\description{
Create a material icon
}
\examples{
material_icon(
 name = "add"
)
}
