% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_profile_card.R
\name{material_profile_card}
\alias{material_profile_card}
\title{Create profile card}
\usage{

  material_profile_card(header_background = "https://images3.alphacoders.com/601/601818.jpg",
  user_img = NULL, user_name = NULL, user_position = NULL,
  user_phone = NULL, user_mail = NULL, date_of_birth = NULL)
}
\arguments{
\item{...}{Slot for any UI element.}

\item{id}{Tag id. Should be unique and correspond to the id provided in material_tabs.}
}
\description{
Create profile card
}
\examples{
material_card_tabs(
 id = "tab_1"
 "Content"
)
}
