% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material_profile_card.R
\name{material_profile_card}
\alias{material_profile_card}
\title{Create profile card}
\usage{

  material_profile_card(header_background = "https://images3.alphacoders.com/601/601818.jpg",
  user_img = NULL, user_name = NULL, user_position = NULL,
  user_phone = NULL, user_mail = NULL, date_of_birth = NULL)
}
\arguments{
\item{header_background}{Header background, if any.}

\item{user_img}{User image, if any.}

\item{user_name}{User name.}

\item{user_position}{User job.}

\item{user_phone}{User phone number, if any.}

\item{user_mail}{User mail adress.}

\item{date_of_birth}{User date of birth.}
}
\description{
Create profile card
}
\examples{
material_profile_card(
 header_background = "https://images3.alphacoders.com/601/601818.jpg",
 user_img = "file:///Users/macdavidgranjon/Desktop/materialize-admin-dashboard-master/images/avatar.jpg",
 user_name = "ROger Waters",
 user_position = "Project Manager",
 user_phone = "+1 (612) 222 8989",
 user_adress = "mail@domain.com"
)
}
